code/durablefunction/requirements.txt
azure-functions~=1.18.0
azure-functions-durable~=1.2.9
pydantic-settings~=2.2.1
pydantic~=2.6.4
azure-identity~=1.15.0
azure-storage-blob~=12.19.1
moviepy~=1.0.3
